Money is always a whole number, in every currency
budget, estimated_cost and cost were decimal(18,2), so forms showed
7,000,000.00. They are now integer columns (old values rounded), with
integer casts and validation, data-money="int" inputs, and Money::format()
never shows decimals.

+++ app/Http/Controllers/ProjectController.php
@@ -131,7 +131,7 @@ class ProjectController extends Controller
             'status' => ['required', Rule::enum(ProjectStatus::class)],
             'start_date' => ['nullable', 'string', $dateRule],
             'end_date' => ['nullable', 'string', $dateRule],
-            'budget' => ['nullable', 'numeric', 'min:0', 'max:9999999999999999'],
+            'budget' => ['nullable', 'integer', 'min:0', 'max:9999999999999999'],
             'currency' => ['required', Rule::in(Money::CURRENCIES)],
             'phone1' => ['nullable', 'string', 'max:30'],
             'phone2' => ['nullable', 'string', 'max:30'],
